financial_report_parser/requirements.txt
pdfplumber==0.10.3
python-dotenv==1.0.0
requests==2.31.0
colorama==0.4.6
